+++ user/src/main/java/com/zenthrex/user/dto/ProfileVerificationDto.java
@@ -35,6 +35,6 @@ public record ProfileVerificationDto(
         LocalDateTime reviewedAt,
 
         @Schema(description = "Reviewed by agent")
-        Long reviewedBy
+        Integer reviewedBy
 ) {
 }
